+++ kernel-d-db/db-mds-sample/src/main/java/io/entframework/kernel/db/mds/example/pojo/request/AutoIncrementRequest.java
@@ -21,7 +21,7 @@ public class AutoIncrementRequest extends BaseRequest {
      */
     @NotNull(message = "Id不能为空", groups = {update.class, delete.class, detail.class, updateStatus.class})
     @ChineseDescription("Id")
-    private Integer id;
+    private Long id;
 
     /**
      * 姓名
@@ -31,5 +31,5 @@ public class AutoIncrementRequest extends BaseRequest {
 
     @NotNull(message = "ID集合不能为空", groups = {batchDelete.class})
     @ChineseDescription("ID集合")
-    private java.util.List<Integer> ids;
+    private java.util.List<Long> ids;
 }
